examples/hotel_app/pyproject.toml
[project]
name = "agent-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
ignore_no_config=true
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.19.0",
    "pydantic[email]>=2.10.4",
    "python-dotenv>=1.0.1",
    "pytest-asyncio>=0.25.2",
    "pydantic-ai[logfire]>=0.0.18",
    "pyjwt>=2.10.1",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
